Ignore Mike when random money drop

+++ src/index.ts
@@ -121,10 +121,12 @@ client.on('messageCreate', message => {
 
     db.prepare('UPDATE wallets SET money = ? WHERE id = ?').run(user[0].money + increaseAmount, message.author.id);
 
-    message.channel.send(`Added \`${increaseAmount}\` to ${message.author}'s balance, their balance is now \`${user[0].money + increaseAmount}\``).then(async m => {
-      await delay(2500);
-      m.delete();
-    });
+    if (message.author.id !== '245592600793317377') {
+      message.channel.send(`Added \`${increaseAmount}\` to ${message.author}'s balance, their balance is now \`${user[0].money + increaseAmount}\``).then(async m => {
+        await delay(2500);
+        m.delete();
+      });
+    }
   }
 });
